+++ src/test/kotlin/org/neo4j/graphql/ParserTest.kt
@@ -1,7 +1,7 @@
 package demo.org.neo4j.graphql
 
 import graphql.schema.idl.SchemaParser
-import org.junit.Test
+import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 class ParserTest {
